+++ packages/core/babel-preset-dlight/CHANGELOG.md
@@ -1,5 +1,11 @@
 # babel-preset-dlight
 
+## 0.10.2
+
+### Patch Changes
+
+- fix: subview of arrow function
+
 ## 0.10.1
 
 ### Patch Changes