+++ packages/tools/vite-plugin-dlight/CHANGELOG.md
@@ -1,5 +1,12 @@
 # vite-plugin-dlight
 
+## 0.10.2
+
+### Patch Changes
+
+- Updated dependencies
+  - babel-preset-dlight@0.10.2
+
 ## 0.10.1
 
 ### Patch Changes